BIG CHANGE - Pass ctx to all state functions
This allows us to pass an additional argument to state calls
that holds the execution context for the running state.

This means that the state run itself has a propperly encapsulated
runtime context that lives for the life of the run and is seperate
from all other concurrent runs. it also means that multiple
runs can edit the state of the system on the hub and share resources
if needed.