Add transparent requisites!!!!!!!!!
Thats RIGHT! Transparent requisites allows for requisites
to be applied to states in the backgroung. So now when you define
a state you can add a variable to the top of the state called TREQ
that refines what requisites will be applied.

There is a BIG TODO though, it does not check for recursive
requisites yet.